Set value of TERM before connecting to server
This fixes an issue when opening menu-based terminal programs, such as vim. If this is not specified, the value of TERM from the ssh client machine will be used, which in my case is xterm-kitty, which is not available on server machines unless I install kitty terminal emulator on them. I do not want to do that, waste of space.